Simplified Notification Vuex store in @codotype/ui
