--- packages/ui/src/modules/notification/store/actions.js
@@ -1,6 +1,0 @@
-
-// actions
-// functions that causes side effects and can involve asynchronous operations.
-// TODO - delete if no actions are required
-export default {
-}
