--- packages/ui/src/modules/notification/store/constants.js
@@ -1,1 +1,0 @@
-// TODO - PROJECT MODULE CONSTANTS HERE
