--- packages/ui/src/modules/notification/store/getters.js
@@ -1,6 +1,0 @@
-
-export default {
-  collection: state => {
-    return state.collection
-  }
-}
