--- packages/ui/src/modules/notification/store/state.js
@@ -1,10 +1,0 @@
-
-// Notification Module State
-export default {
-  collection: [
-    // Notification examples:
-    // { dismissible: true, strong: 'Warning', message: 'Something went wrong.', context: 'warning' },
-    // { dismissible: true, strong: 'Info', message: 'Something went wrong.', context: 'info' },
-    // { dismissible: true, strong: 'Danger!', message: 'Something went wrong.', context: 'danger' }
-  ]
-}
